Renamed script to "pipeline", since it isn't specific to NetCDFs anymore.
